Styling
The typography, colors, and layout has been established. Some layout has been applied for small screens, but the transitions to larger screens need to be worked on.

The header and footer still need to be styled, the layout has to be worked on to adjust to different screen sizes, the menu still needs to be styled and icons still need to be added.