Style header and footer
Style the header `h1` to break up the long text, playing with size,
alignment, and decoration, to try to help make it digestible and "pop".

Consolidate some heading styles where we can leverage the cascade.

Generally center the footer elements, with various degrees of
sophistication for control.

Use `text-decoration` to implement the underline instead of
`border-bottom`. This achieves the same effect, and has two main benefits:

1. Links will wrap within the word, helping responsive behavior.
2. The underline will draw behind low dangling glyph segments, like
`g` or `j`, instead of appearing below the entire glyph, which better
reflects how we draw letters on lines.